requirements.txt
tensorflow-gpu==2.2.0-rc2
tfa-nightly==0.10.0.dev20200502100543
mock
tqdm
numpy
miniimagenettools
pandas
tensorflow_hub
#  Just for plots and graphs
matplotlib
seaborn
# for face recognition
# face_recognition
